Refactored parts of ITCV. New enum names. Fixed bug where duplicate registering as listener. Removed old UI logics.

+++ victorious/VictoriousIOSSDK/AcknowledgeAlertRequest.swift
@@ -11,16 +11,16 @@ import Foundation
 /// Marks an alert as seen by the user, which will remove it from any further response payloads.
 public struct AcknowledgeAlertRequest: RequestType {
 
-    public let alertID: Int
+    public let alertID: Double
     
     public var urlRequest: NSURLRequest {
         let urlRequest = NSMutableURLRequest(URL: NSURL(string: "/api/alert/acknowledge")!)
-        let params = ["alert_id": String(alertID)]
+        let params = ["alert_id": String(Int(alertID))]
         urlRequest.vsdk_addURLEncodedFormPost(params)
         return urlRequest
     }
     
-    public init(alertID: Int) {
+    public init(alertID: Double) {
         self.alertID = alertID
     }
 }
